+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.3.0] - 2021-03-29
+
+### Fixed
+
+- Models relationships;
+
+### Added
+
+- Notebooks controllers and routes;
+
 ## [0.2.0] - 2021-03-28
 
 ### Changed